FASTAPI/fastapi_amisr14/requirements.txt
absl-py==2.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apturl==0.5.2
asgiref==3.5.0
astunparse==1.6.3
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.3.0
fastapi==0.121.0
fasteners==0.14.1
flatbuffers==25.2.10
fonttools==4.60.1
future==0.18.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.13.0
h5py==3.13.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keras==3.9.0
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mypy==1.18.2
mypy_extensions==1.1.0
namex==0.0.8
netifaces==0.11.0
numpy==2.1.3
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.5.82
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
paramiko==2.9.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
protobuf==5.29.3
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==13.9.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
ssh-import-id==5.11
starlette==0.49.3
systemd-python==234
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
texttable==1.6.4
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
